+++ backend/uclapi/workspaces/occupeye/exceptions.py
@@ -1,0 +1,8 @@
+class BadOccupEyeRequest(Exception):
+    """
+    Custom exception for when any CadCap API request fails.
+    This should only be raised by endpoints that could contain
+    user-entered data, so that we can give them an error telling
+    them that the data they gave us is bad.
+    """
+    pass
